Fix: Content-Length type from int to size_t

+++ src/HTTPRequest.cpp
@@ -153,11 +153,11 @@ const char* httplib::HTTPRequest::getBodyPointer()
 	return body;
 }
 
-int httplib::HTTPRequest::getBodyLen(){
+size_t httplib::HTTPRequest::getBodyLen(){
 	return body_len;
 }
 
-void httplib::HTTPRequest::setBody(const char *pbody, const int len){
+void httplib::HTTPRequest::setBody(const char *pbody, const size_t len){
 	if (pbody == nullptr){
 		free(body);
 		body = nullptr;
